Initial commit via ai
ai help up to this point. I realized I had to get cards done first as the primer so there will be a lot of rework on this. I also got the DO droplet up and running and it has a landing .html but I believe set up through apache... so that will need work also.